improve profile switching
The keyboard is used with Mac and Windows. For both, different layer
stacks exist (base, vim, symbols and numbers).

When using the profile for Windows (4), I necessarily only want to do
use that profile with the layer 5 active. Analogously, when using the
profile for Mac (1) I only want use that profile with layer 0 active.

To combine profile switching and the switch of layer stacks (base1 to
base2 or base2 to base1), macros are created.

The reasoning holds analogously for profiles 2 (iPad) and 3 (iPhone),
hence macros for those profiles have been created as well.

A concern was whether profile clearing with BT_CLR would still work, but
I think it should since the key will still be pressed while in some
bluetooth profile, only that the switch to that profile was followed up
by a base layer switch.