relocate HRMs to BRMs
HRM - home row mods to BRMS - bottom row mods

why? because for one on the right split side where hjkl are used constantly.
Hence, the delay in the key press, which is only triggered upon key
release (before tapping term has expired), seems to be an issue.

and for symmetry reasons, the keys have been relocated on the left split
side as well.