move BRMs on left hand to HRMs on left hand
I transitioned in b978233 (and 546466f) to BRMs because when in Vim
hjkl are used a lot, and the few milliseconds between key down and key
up do add up. With HRMs on right hand, the hjkl get only fired on key up
since at key down the duration is not clear yet.

I moved to BRMs on both halves for symmetry reasons, but came to realise
HRMs on the left side would be more comfortable and should not clash with
other functionality.